requirements.txt
alembic==1.16.1
blinker==1.9.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.3
Flask==3.1.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
logger==1.4
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.1.3
